Improve passage interface
Apply themeSpecial to passage interface table headers.

Limit the width of passage interface table columns. The minimum
terminal width is assumed to be 72 columns wide. Four of these columns
are lost to border and border padding. The UID is only allowed to be
seven columns long. The date is always ten columns wide. There is a
single column between table columns for readability, which leaves 49
columns for the passage name.