Replace HasDefault class with undefined value
Remove the HasDefault class, which provided defaultValue as the
serialized value of the Transient type. Replace it with the undefined
value. This change takes advantage of Haskell's lazy evaluation to
simplify code. Documentation on the use of Transient should be added in
the future, either to haskell-typist or to flat.