Rewrite interfaces to work with FocusedList
Rewrite the Interface type to work with FocusedList instead of naked
indices. This change greatly improves the readability and
maintainability of the entire application by coupling the list and the
focus.

Move the passage sorting process from the Interface module to the
Passage module. This is a more natural location and reduces the design
load of the Interface module.

Draw up to three passages above and below the focused passage.